+++ server/src/main/java/DatabaseAccess/FFUserDao.java
@@ -1,4 +1,20 @@
 package DatabaseAccess;
 
+import com.example.shared.model.User;
+
 public class FFUserDao implements IUserDao {
+    @Override
+    public void registerUser(User user) {
+
+    }
+
+    @Override
+    public User getUser(String username) {
+        return null;
+    }
+
+    @Override
+    public void clear() {
+
+    }
 }
